Alias artifactory_unmanaged_user to same code as artifactory_user
This remove the last SDKv2 code from user resources, and simplify maintenance.

Make internal_password_disabled field into pointer and only send to API if it is set with a value.

Do not send 'password' in API request if internal_password_disabled is true
