+++ pkg/artifactory/resource/user/resource_artifactory_unmanaged_user_test.go
@@ -50,9 +50,9 @@ func TestAccUnmanagedUser_PasswordNotChangeWhenOtherAttributesChangeGH340(t *tes
 	`, params)
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { acctest.PreCheck(t) },
-		ProviderFactories: acctest.ProviderFactories,
-		CheckDestroy:      testAccCheckUserDestroy(fqrn),
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
 		Steps: []resource.TestStep{
 			{
 				Config: userInitial,
@@ -96,9 +96,9 @@ func TestAccUnmanagedUser_basic(t *testing.T) {
 	fqrn := fmt.Sprintf("artifactory_unmanaged_user.%s", name)
 	username := fmt.Sprintf("dummy_user%d", id)
 	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { acctest.PreCheck(t) },
-		ProviderFactories: acctest.ProviderFactories,
-		CheckDestroy:      testAccCheckUserDestroy(fqrn),
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
 		Steps: []resource.TestStep{
 			{
 				Config: fmt.Sprintf(userBasic, name, username, id),
@@ -125,6 +125,7 @@ func TestAccUnmanagedUser_ShouldCreateUpdateWithoutPassword(t *testing.T) {
 			name  	= "%s"
 			email 	= "dummy_user%d@a.com"
 			profile_updatable = true
+			disable_ui_access = false
 		}
 	`
 
@@ -132,7 +133,8 @@ func TestAccUnmanagedUser_ShouldCreateUpdateWithoutPassword(t *testing.T) {
 		resource "artifactory_unmanaged_user" "%s" {
 			name  	= "%s"
 			email 	= "dummy_user%d@a.com"
-			profile_updatable = false
+			profile_updatable = true
+			disable_ui_access = true
 		}
 	`
 
@@ -142,9 +144,9 @@ func TestAccUnmanagedUser_ShouldCreateUpdateWithoutPassword(t *testing.T) {
 	username := fmt.Sprintf("dummy_user%d", id)
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { acctest.PreCheck(t) },
-		ProviderFactories: acctest.ProviderFactories,
-		CheckDestroy:      testAccCheckUserDestroy(fqrn),
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
 		Steps: []resource.TestStep{
 			{
 				Config: fmt.Sprintf(config, name, username, id),
@@ -152,6 +154,7 @@ func TestAccUnmanagedUser_ShouldCreateUpdateWithoutPassword(t *testing.T) {
 					resource.TestCheckResourceAttr(fqrn, "name", username),
 					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
 					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "disable_ui_access", "false"),
 					resource.TestCheckNoResourceAttr(fqrn, "password"),
 					resource.TestCheckNoResourceAttr(fqrn, "groups"),
 				),
@@ -161,17 +164,128 @@ func TestAccUnmanagedUser_ShouldCreateUpdateWithoutPassword(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(fqrn, "name", username),
 					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
-					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "false"),
+					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "disable_ui_access", "true"),
+					resource.TestCheckNoResourceAttr(fqrn, "password"),
+					resource.TestCheckResourceAttr(fqrn, "groups.#", "0"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccUnmanagedUser_internal_password_disabled_changed_error(t *testing.T) {
+	const config = `
+		resource "artifactory_unmanaged_user" "%s" {
+			name  	= "%s"
+			email 	= "dummy_user%d@a.com"
+			profile_updatable = true
+			disable_ui_access = false
+			internal_password_disabled = true
+		}
+	`
+
+	const updatedConfig = `
+		resource "artifactory_unmanaged_user" "%s" {
+			name  	= "%s"
+			email 	= "dummy_user%d@a.com"
+			disable_ui_access = true
+			internal_password_disabled = false
+		}
+	`
+
+	id := testutil.RandomInt()
+	name := fmt.Sprintf("foobar-%d", id)
+	fqrn := fmt.Sprintf("artifactory_unmanaged_user.%s", name)
+	username := fmt.Sprintf("dummy_user%d", id)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(config, name, username, id),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", username),
+					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
+					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "internal_password_disabled", "true"),
 					resource.TestCheckNoResourceAttr(fqrn, "password"),
 					resource.TestCheckNoResourceAttr(fqrn, "groups"),
 				),
 			},
 			{
-				ResourceName:            fqrn,
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateCheck:        validator.CheckImportState(username, "name"),
-				ImportStateVerifyIgnore: []string{"password"}, // password is never returned via the API, so it cannot be "imported"
+				Config:      fmt.Sprintf(updatedConfig, name, username, id),
+				ExpectError: regexp.MustCompile(`Password must be set when internal_password_disabled is changed to 'false'`),
+			},
+		},
+	})
+}
+
+func TestAccUnmanagedUser_internal_password_disabled_changed_with_password(t *testing.T) {
+	const config = `
+		resource "artifactory_unmanaged_user" "%s" {
+			name  	= "%s"
+			email 	= "dummy_user%d@a.com"
+			profile_updatable = true
+			disable_ui_access = false
+			internal_password_disabled = true
+		}
+	`
+
+	const updatedConfig = `
+		resource "artifactory_unmanaged_user" "%s" {
+			name  	= "%s"
+			email 	= "dummy_user%d@a.com"
+			password = "Password!123"
+			disable_ui_access = true
+			internal_password_disabled = false
+		}
+	`
+
+	id := testutil.RandomInt()
+	name := fmt.Sprintf("foobar-%d", id)
+	fqrn := fmt.Sprintf("artifactory_unmanaged_user.%s", name)
+	username := fmt.Sprintf("dummy_user%d", id)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(config, name, username, id),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", username),
+					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
+					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "internal_password_disabled", "true"),
+					resource.TestCheckNoResourceAttr(fqrn, "password"),
+					resource.TestCheckNoResourceAttr(fqrn, "groups"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(updatedConfig, name, username, id),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", username),
+					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
+					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "internal_password_disabled", "false"),
+					resource.TestCheckResourceAttr(fqrn, "password", "Password!123"),
+					resource.TestCheckResourceAttr(fqrn, "groups.#", "0"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(config, name, username, id),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", username),
+					resource.TestCheckResourceAttr(fqrn, "email", fmt.Sprintf("dummy_user%d@a.com", id)),
+					resource.TestCheckResourceAttr(fqrn, "profile_updatable", "true"),
+					resource.TestCheckResourceAttr(fqrn, "internal_password_disabled", "true"),
+					resource.TestCheckNoResourceAttr(fqrn, "password"),
+					resource.TestCheckResourceAttr(fqrn, "groups.#", "0"),
+				),
 			},
 		},
 	})
@@ -245,7 +359,7 @@ func TestAccUnmanagedUser_full(t *testing.T) {
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
-		ProtoV6ProviderFactories: acctest.ProtoV6MuxProviderFactories,
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
 		CheckDestroy:             testAccCheckUserDestroy(fqrn),
 		Steps: []resource.TestStep{
 			{
@@ -302,7 +416,7 @@ func TestAccUnmanagedUser_invalidName(t *testing.T) {
 		username   string
 		errorRegex string
 	}{
-		{"Empty", "", `.*expected "name" to not be an empty string.*`},
+		{"Empty", "", `.*Attribute name string length must be at least 1, got: 0.*`},
 		{"Uppercase", "test_user_Uppercase", `.*may contain lowercase letters, numbers and symbols: '.-_@'.*`},
 		{"Symbols", "test_user_!", `.*may contain lowercase letters, numbers and symbols: '.-_@'.*`},
 	}
@@ -323,9 +437,9 @@ func testAccUnmanagedUserInvalidName(username, errorRegex string) func(t *testin
 		id, fqrn, name := testutil.MkNames("test-", "artifactory_unmanaged_user")
 
 		resource.Test(t, resource.TestCase{
-			PreCheck:          func() { acctest.PreCheck(t) },
-			ProviderFactories: acctest.ProviderFactories,
-			CheckDestroy:      testAccCheckUserDestroy(fqrn),
+			PreCheck:                 func() { acctest.PreCheck(t) },
+			ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+			CheckDestroy:             testAccCheckUserDestroy(fqrn),
 			Steps: []resource.TestStep{
 				{
 					Config:      fmt.Sprintf(userNoGroups, name, username, id),
@@ -348,9 +462,9 @@ func TestAccUnmanagedUser_EmptyGroups(t *testing.T) {
 	id, fqrn, name := testutil.MkNames("foobar-", "artifactory_unmanaged_user")
 	username := fmt.Sprintf("dummy_user%d", id)
 	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { acctest.PreCheck(t) },
-		ProviderFactories: acctest.ProviderFactories,
-		CheckDestroy:      testAccCheckUserDestroy(fqrn),
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckUserDestroy(fqrn),
 		Steps: []resource.TestStep{
 			{
 				Config: fmt.Sprintf(userEmptyGroups, name, username, id),
