+++ pkg/artifactory/resource/user/resource_artifactory_user.go
@@ -5,8 +5,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/samber/lo"
 )
 
@@ -30,9 +28,6 @@ func (r *ArtifactoryUserResource) Schema(ctx context.Context, req resource.Schem
 			Optional:  true,
 			Sensitive: true,
 			Computed:  true,
-			PlanModifiers: []planmodifier.String{
-				stringplanmodifier.UseStateForUnknown(),
-			},
 		},
 	}
